Add Jest Snapshot and configure CircleCi parallelism
