+++ Example/.babelrc
@@ -1,3 +1,3 @@
 {
-  presets: ["react-native"]
+  "presets": ["react-native"]
 }
