+++ Example/__tests__/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+    "env": {
+        "jest": true
+    }
+}
