+++ Example/package.json
@@ -4,17 +4,26 @@
   "private": true,
   "scripts": {
     "start": "node_modules/react-native/packager/packager.sh",
-    "sync-rnrf": "rm -rf ./node_modules/react-native-router-flux; sane '/usr/bin/rsync -v -a --exclude .git --exclude Example --exclude node_modules ../ ./node_modules/react-native-router-flux/' .. --glob='{**/*.json,**/*.js}'"
+    "sync-rnrf": "rm -rf ./node_modules/react-native-router-flux; sane '/usr/bin/rsync -v -a --exclude .git --exclude Example --exclude node_modules ../ ./node_modules/react-native-router-flux/' .. --glob='{**/*.json,**/*.js}'",
+    "jest": "jest"
   },
   "dependencies": {
     "clamp": "^1.0.1",
-    "react": "^15.1.0",
-    "react-native": "^0.27.0",
+    "react": "15.3.0",
+    "react-native": "0.30.0",
     "react-native-button": "github:ide/react-native-button",
     "react-native-drawer": "^2.2.2",
     "react-native-router-flux": "file:../"
   },
   "devDependencies": {
+    "babel-jest": "^14.1.0",
+    "babel-preset-react-native": "^1.9.0",
+    "jest": "^14.1.0",
+    "jest-react-native": "^14.1.1",
+    "react-test-renderer": "^15.3.0",
     "sane": "^1.3.4"
+  },
+  "jest": {
+    "preset": "jest-react-native"
   }
 }
